refactor: 할인 금액 객체 분리
